Basic array list to minic some functionality of collections.
This array list to allow for a non-intrusive method of storing a
collection of data, which may or may not be a model, but will still
allow for event driven programming, and also allow for a model to be
bound to more than 1 array list. In contrast to collection in which a
model can only belong to one, an array list will allow for a model to
exist in unlimited number of arraylists, and even exist in the same
array list. Ensuring no duplicates is a matter for the developer.